feat(PocEntryPoint.cpp, PocGraphModule.cpp, PocGraphModule.hpp): add POC graph module
Add a proof of concept (POC) graph module to the project. The POC graph module
is a simple module that writes the rendered image to a PPM file. The POC graph
module consists of three files: PocEntryPoint.cpp, PocGraphModule.cpp, and
PocGraphModule.hpp. The PocEntryPoint.cpp file contains the entry point for the
graph module, which returns an instance of the PocGraphModule class. The
PocGraphModule.cpp file contains the implementation of the PocGraphModule
class, which writes the rendered image to a PPM file. The PocGraphModule.hpp
file contains the declaration of the PocGraphModule class.