feat(Cones): add Cone primitive
Add Cone primitive to the raytracer. Cone is a subclass of APrimitive and
implements the scale and intersect methods. Cone is created with a
position, radius and color. The intersect method calculates the
intersection of a ray with the cone and returns an optional unique_ptr
to an Intersection object. The scale method scales the radius of the
cone. The Cone class is defined in Cone.hpp and implemented in Cone.cpp.
The Cone plugin is created in ConesEntryPoint.cpp and exports the
create, destroy and getName functions.